Add support for creating layered war files with Gradle
See gh-22195
