+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/main/java/org/springframework/boot/gradle/tasks/bundling/LayeredSpec.java
@@ -41,7 +41,7 @@ import org.springframework.boot.loader.tools.layer.LibraryContentFilter;
 import org.springframework.util.Assert;
 
 /**
- * Encapsulates the configuration for a layered jar.
+ * Encapsulates the configuration for a layered archive.
  *
  * @author Madhura Bhave
  * @author Scott Frederick
@@ -65,7 +65,7 @@ public class LayeredSpec {
 
 	/**
 	 * Returns whether the layer tools should be included as a dependency in the layered
-	 * jar.
+	 * archive.
 	 * @return whether the layer tools should be included
 	 */
 	@Input
@@ -74,7 +74,8 @@ public class LayeredSpec {
 	}
 
 	/**
-	 * Sets whether the layer tools should be included as a dependency in the layered jar.
+	 * Sets whether the layer tools should be included as a dependency in the layered
+	 * archive.
 	 * @param includeLayerTools {@code true} if the layer tools should be included,
 	 * otherwise {@code false}
 	 */
@@ -83,7 +84,7 @@ public class LayeredSpec {
 	}
 
 	/**
-	 * Returns whether the layers.idx should be included in the jar.
+	 * Returns whether the layers.idx should be included in the archive.
 	 * @return whether the layers.idx should be included
 	 */
 	@Input
@@ -92,8 +93,8 @@ public class LayeredSpec {
 	}
 
 	/**
-	 * Sets whether the layers.idx should be included in the jar.
-	 * @param enabled {@code true} layers.idx should be included in the jar, otherwise
+	 * Sets whether the layers.idx should be included in the archive.
+	 * @param enabled {@code true} layers.idx should be included in the archive, otherwise
 	 * {@code false}
 	 */
 	public void setEnabled(boolean enabled) {
@@ -171,7 +172,8 @@ public class LayeredSpec {
 	}
 
 	/**
-	 * Returns the order of the layers in the jar from least to most frequently changing.
+	 * Returns the order of the layers in the archive from least to most frequently
+	 * changing.
 	 * @return the layer order
 	 */
 	@Input
@@ -180,7 +182,7 @@ public class LayeredSpec {
 	}
 
 	/**
-	 * Sets to order of the layers in the jar from least to most frequently changing.
+	 * Sets the order of the layers in the archive from least to most frequently changing.
 	 * @param layerOrder the layer order
 	 */
 	public void setLayerOrder(String... layerOrder) {
@@ -188,7 +190,7 @@ public class LayeredSpec {
 	}
 
 	/**
-	 * Sets to order of the layers in the jar from least to most frequently changing.
+	 * Sets the order of the layers in the archive from least to most frequently changing.
 	 * @param layerOrder the layer order
 	 */
 	public void setLayerOrder(List<String> layerOrder) {
