+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/test/java/org/springframework/boot/gradle/tasks/bundling/AbstractBootArchiveTests.java
@@ -16,10 +16,12 @@
 
 package org.springframework.boot.gradle.tasks.bundling;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.nio.file.Files;
 import java.nio.file.StandardOpenOption;
 import java.nio.file.attribute.PosixFilePermission;
@@ -27,19 +29,35 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Enumeration;
 import java.util.HashMap;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.TreeSet;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
 import java.util.jar.JarOutputStream;
 import java.util.jar.Manifest;
+import java.util.stream.Collectors;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
 import org.apache.commons.compress.archivers.zip.ZipArchiveEntry;
 import org.apache.commons.compress.archivers.zip.ZipFile;
+import org.gradle.api.Action;
+import org.gradle.api.DomainObjectSet;
 import org.gradle.api.Project;
+import org.gradle.api.artifacts.Configuration;
+import org.gradle.api.artifacts.DependencySet;
+import org.gradle.api.artifacts.ModuleVersionIdentifier;
+import org.gradle.api.artifacts.ProjectDependency;
+import org.gradle.api.artifacts.ResolvableDependencies;
+import org.gradle.api.artifacts.ResolvedArtifact;
+import org.gradle.api.artifacts.ResolvedConfiguration;
+import org.gradle.api.artifacts.ResolvedModuleVersion;
+import org.gradle.api.artifacts.component.ComponentArtifactIdentifier;
+import org.gradle.api.artifacts.component.ModuleComponentIdentifier;
+import org.gradle.api.artifacts.component.ProjectComponentIdentifier;
 import org.gradle.api.internal.file.archive.ZipCopyAction;
 import org.gradle.api.tasks.bundling.AbstractArchiveTask;
 import org.gradle.api.tasks.bundling.Jar;
@@ -49,8 +67,13 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import org.springframework.boot.loader.tools.DefaultLaunchScript;
+import org.springframework.boot.loader.tools.JarModeLibrary;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.willAnswer;
+import static org.mockito.Mockito.mock;
 
 /**
  * Abstract base class for testing {@link BootArchive} implementations.
@@ -72,15 +95,19 @@ abstract class AbstractBootArchiveTests<T extends Jar & BootArchive> {
 
 	private final String classesPath;
 
+	private final String indexPath;
+
 	private Project project;
 
 	private T task;
 
-	protected AbstractBootArchiveTests(Class<T> taskClass, String launcherClass, String libPath, String classesPath) {
+	protected AbstractBootArchiveTests(Class<T> taskClass, String launcherClass, String libPath, String classesPath,
+			String indexPath) {
 		this.taskClass = taskClass;
 		this.launcherClass = launcherClass;
 		this.libPath = libPath;
 		this.classesPath = classesPath;
+		this.indexPath = indexPath;
 	}
 
 	@BeforeEach
@@ -407,6 +434,145 @@ abstract class AbstractBootArchiveTests<T extends Jar & BootArchive> {
 				this.libPath + "third-library.jar");
 	}
 
+	@Test
+	void archiveShouldBeLayeredByDefault() throws IOException {
+		addContent();
+		executeTask();
+		try (JarFile jarFile = new JarFile(this.task.getArchiveFile().get().getAsFile())) {
+			assertThat(jarFile.getManifest().getMainAttributes().getValue("Spring-Boot-Classes"))
+					.isEqualTo(this.classesPath);
+			assertThat(jarFile.getManifest().getMainAttributes().getValue("Spring-Boot-Lib")).isEqualTo(this.libPath);
+			assertThat(jarFile.getManifest().getMainAttributes().getValue("Spring-Boot-Layers-Index"))
+					.isEqualTo(this.indexPath + "layers.idx");
+			assertThat(getEntryNames(jarFile)).contains(this.libPath + JarModeLibrary.LAYER_TOOLS.getName());
+		}
+	}
+
+	@Test
+	void jarWhenLayersDisabledShouldNotContainLayersIndex() throws IOException {
+		List<String> entryNames = getEntryNames(createLayeredJar((configuration) -> configuration.setEnabled(false)));
+		assertThat(entryNames).doesNotContain(this.indexPath + "layers.idx");
+	}
+
+	@Test
+	void whenJarIsLayeredThenManifestContainsEntryForLayersIndexInPlaceOfClassesAndLib() throws IOException {
+		try (JarFile jarFile = new JarFile(createLayeredJar())) {
+			assertThat(jarFile.getManifest().getMainAttributes().getValue("Spring-Boot-Classes"))
+					.isEqualTo(this.classesPath);
+			assertThat(jarFile.getManifest().getMainAttributes().getValue("Spring-Boot-Lib")).isEqualTo(this.libPath);
+			assertThat(jarFile.getManifest().getMainAttributes().getValue("Spring-Boot-Layers-Index"))
+					.isEqualTo(this.indexPath + "layers.idx");
+		}
+	}
+
+	@Test
+	void whenJarIsLayeredThenLayersIndexIsPresentAndCorrect() throws IOException {
+		try (JarFile jarFile = new JarFile(createLayeredJar())) {
+			List<String> entryNames = getEntryNames(jarFile);
+			assertThat(entryNames).contains(this.libPath + "first-library.jar", this.libPath + "second-library.jar",
+					this.libPath + "third-library-SNAPSHOT.jar", this.libPath + "first-project-library.jar",
+					this.libPath + "second-project-library-SNAPSHOT.jar",
+					this.classesPath + "com/example/Application.class", this.classesPath + "application.properties",
+					this.classesPath + "static/test.css");
+			List<String> index = entryLines(jarFile, this.indexPath + "layers.idx");
+			assertThat(getLayerNames(index)).containsExactly("dependencies", "spring-boot-loader",
+					"snapshot-dependencies", "application");
+			String layerToolsJar = this.libPath + JarModeLibrary.LAYER_TOOLS.getName();
+			List<String> expected = new ArrayList<>();
+			expected.add("- \"dependencies\":");
+			expected.add("  - \"" + this.libPath + "first-library.jar\"");
+			expected.add("  - \"" + this.libPath + "first-project-library.jar\"");
+			expected.add("  - \"" + this.libPath + "second-library.jar\"");
+			if (!layerToolsJar.contains("SNAPSHOT")) {
+				expected.add("  - \"" + layerToolsJar + "\"");
+			}
+			expected.add("- \"spring-boot-loader\":");
+			expected.add("  - \"org/\"");
+			expected.add("- \"snapshot-dependencies\":");
+			expected.add("  - \"" + this.libPath + "second-project-library-SNAPSHOT.jar\"");
+			if (layerToolsJar.contains("SNAPSHOT")) {
+				expected.add("  - \"" + layerToolsJar + "\"");
+			}
+			expected.add("  - \"" + this.libPath + "third-library-SNAPSHOT.jar\"");
+			expected.add("- \"application\":");
+			Set<String> applicationContents = new TreeSet<>();
+			applicationContents.add("  - \"" + this.classesPath + "\"");
+			if (archiveHasClasspathIndex()) {
+				applicationContents.add("  - \"" + this.indexPath + "classpath.idx\"");
+			}
+			applicationContents.add("  - \"" + this.indexPath + "layers.idx\"");
+			applicationContents.add("  - \"META-INF/\"");
+			expected.addAll(applicationContents);
+			assertThat(index).containsExactlyElementsOf(expected);
+		}
+	}
+
+	@Test
+	void whenJarIsLayeredWithCustomStrategiesThenLayersIndexIsPresentAndCorrect() throws IOException {
+		File jar = createLayeredJar((layered) -> {
+			layered.application((application) -> {
+				application.intoLayer("resources", (spec) -> spec.include("static/**"));
+				application.intoLayer("application");
+			});
+			layered.dependencies((dependencies) -> {
+				dependencies.intoLayer("my-snapshot-deps", (spec) -> spec.include("com.example:*:*.SNAPSHOT"));
+				dependencies.intoLayer("my-internal-deps", (spec) -> spec.include("com.example:*:*"));
+				dependencies.intoLayer("my-deps");
+			});
+			layered.setLayerOrder("my-deps", "my-internal-deps", "my-snapshot-deps", "resources", "application");
+		});
+		try (JarFile jarFile = new JarFile(jar)) {
+			List<String> entryNames = getEntryNames(jar);
+			assertThat(entryNames).contains(this.libPath + "first-library.jar", this.libPath + "second-library.jar",
+					this.libPath + "third-library-SNAPSHOT.jar", this.libPath + "first-project-library.jar",
+					this.libPath + "second-project-library-SNAPSHOT.jar",
+					this.classesPath + "com/example/Application.class", this.classesPath + "application.properties",
+					this.classesPath + "static/test.css");
+			List<String> index = entryLines(jarFile, this.indexPath + "layers.idx");
+			assertThat(getLayerNames(index)).containsExactly("my-deps", "my-internal-deps", "my-snapshot-deps",
+					"resources", "application");
+			String layerToolsJar = this.libPath + JarModeLibrary.LAYER_TOOLS.getName();
+			List<String> expected = new ArrayList<>();
+			expected.add("- \"my-deps\":");
+			expected.add("  - \"" + layerToolsJar + "\"");
+			expected.add("- \"my-internal-deps\":");
+			expected.add("  - \"" + this.libPath + "first-library.jar\"");
+			expected.add("  - \"" + this.libPath + "first-project-library.jar\"");
+			expected.add("  - \"" + this.libPath + "second-library.jar\"");
+			expected.add("- \"my-snapshot-deps\":");
+			expected.add("  - \"" + this.libPath + "second-project-library-SNAPSHOT.jar\"");
+			expected.add("  - \"" + this.libPath + "third-library-SNAPSHOT.jar\"");
+			expected.add("- \"resources\":");
+			expected.add("  - \"" + this.classesPath + "static/\"");
+			expected.add("- \"application\":");
+			Set<String> applicationContents = new TreeSet<>();
+			applicationContents.add("  - \"" + this.classesPath + "application.properties\"");
+			applicationContents.add("  - \"" + this.classesPath + "com/\"");
+			if (archiveHasClasspathIndex()) {
+				applicationContents.add("  - \"" + this.indexPath + "classpath.idx\"");
+			}
+			applicationContents.add("  - \"" + this.indexPath + "layers.idx\"");
+			applicationContents.add("  - \"META-INF/\"");
+			applicationContents.add("  - \"org/\"");
+			expected.addAll(applicationContents);
+			assertThat(index).containsExactlyElementsOf(expected);
+		}
+	}
+
+	@Test
+	void whenArchiveIsLayeredThenLayerToolsAreAddedToTheJar() throws IOException {
+		List<String> entryNames = getEntryNames(createLayeredJar());
+		assertThat(entryNames).contains(this.libPath + JarModeLibrary.LAYER_TOOLS.getName());
+	}
+
+	@Test
+	void whenArchiveIsLayeredAndIncludeLayerToolsIsFalseThenLayerToolsAreNotAddedToTheJar() throws IOException {
+		List<String> entryNames = getEntryNames(
+				createLayeredJar((configuration) -> configuration.setIncludeLayerTools(false)));
+		assertThat(entryNames)
+				.doesNotContain(this.indexPath + "layers/dependencies/lib/spring-boot-jarmode-layertools.jar");
+	}
+
 	protected File jarFile(String name) throws IOException {
 		File file = newFile(name);
 		try (JarOutputStream jar = new JarOutputStream(new FileOutputStream(file))) {
@@ -451,6 +617,120 @@ abstract class AbstractBootArchiveTests<T extends Jar & BootArchive> {
 		File file = new File(this.temp, name);
 		file.createNewFile();
 		return file;
+	}
+
+	File createLayeredJar() throws IOException {
+		return createLayeredJar((spec) -> {
+		});
+	}
+
+	File createLayeredJar(Action<LayeredSpec> action) throws IOException {
+		applyLayered(action);
+		addContent();
+		executeTask();
+		return getTask().getArchiveFile().get().getAsFile();
+	}
+
+	abstract void applyLayered(Action<LayeredSpec> action);
+
+	boolean archiveHasClasspathIndex() {
+		return true;
+	}
+
+	@SuppressWarnings("unchecked")
+	void addContent() throws IOException {
+		this.task.getMainClass().set("com.example.Main");
+		File classesJavaMain = new File(this.temp, "classes/java/main");
+		File applicationClass = new File(classesJavaMain, "com/example/Application.class");
+		applicationClass.getParentFile().mkdirs();
+		applicationClass.createNewFile();
+		File resourcesMain = new File(this.temp, "resources/main");
+		File applicationProperties = new File(resourcesMain, "application.properties");
+		applicationProperties.getParentFile().mkdirs();
+		applicationProperties.createNewFile();
+		File staticResources = new File(resourcesMain, "static");
+		staticResources.mkdir();
+		File css = new File(staticResources, "test.css");
+		css.createNewFile();
+		this.task.classpath(classesJavaMain, resourcesMain, jarFile("first-library.jar"), jarFile("second-library.jar"),
+				jarFile("third-library-SNAPSHOT.jar"), jarFile("first-project-library.jar"),
+				jarFile("second-project-library-SNAPSHOT.jar"));
+		Set<ResolvedArtifact> artifacts = new LinkedHashSet<>();
+		artifacts.add(mockLibraryArtifact("first-library.jar", "com.example", "first-library", "1.0.0"));
+		artifacts.add(mockLibraryArtifact("second-library.jar", "com.example", "second-library", "1.0.0"));
+		artifacts.add(
+				mockLibraryArtifact("third-library-SNAPSHOT.jar", "com.example", "third-library", "1.0.0.SNAPSHOT"));
+		artifacts
+				.add(mockProjectArtifact("first-project-library.jar", "com.example", "first-project-library", "1.0.0"));
+		artifacts.add(mockProjectArtifact("second-project-library-SNAPSHOT.jar", "com.example",
+				"second-project-library", "1.0.0.SNAPSHOT"));
+		ResolvedConfiguration resolvedConfiguration = mock(ResolvedConfiguration.class);
+		given(resolvedConfiguration.getResolvedArtifacts()).willReturn(artifacts);
+		Configuration configuration = mock(Configuration.class);
+		given(configuration.getResolvedConfiguration()).willReturn(resolvedConfiguration);
+		ResolvableDependencies resolvableDependencies = mock(ResolvableDependencies.class);
+		given(configuration.getIncoming()).willReturn(resolvableDependencies);
+		DependencySet dependencies = mock(DependencySet.class);
+		DomainObjectSet<ProjectDependency> projectDependencies = mock(DomainObjectSet.class);
+		given(dependencies.withType(ProjectDependency.class)).willReturn(projectDependencies);
+		given(configuration.getAllDependencies()).willReturn(dependencies);
+		willAnswer((invocation) -> {
+			invocation.getArgument(0, Action.class).execute(resolvableDependencies);
+			return null;
+		}).given(resolvableDependencies).afterResolve(any(Action.class));
+		given(configuration.getIncoming()).willReturn(resolvableDependencies);
+		populateResolvedDependencies(configuration);
+	}
+
+	abstract void populateResolvedDependencies(Configuration configuration);
+
+	private ResolvedArtifact mockLibraryArtifact(String fileName, String group, String module, String version) {
+		ModuleComponentIdentifier moduleComponentIdentifier = mock(ModuleComponentIdentifier.class);
+		ComponentArtifactIdentifier libraryArtifactId = mock(ComponentArtifactIdentifier.class);
+		given(libraryArtifactId.getComponentIdentifier()).willReturn(moduleComponentIdentifier);
+		ResolvedArtifact libraryArtifact = mockArtifact(fileName, group, module, version);
+		given(libraryArtifact.getId()).willReturn(libraryArtifactId);
+		return libraryArtifact;
+	}
+
+	private ResolvedArtifact mockProjectArtifact(String fileName, String group, String module, String version) {
+		ProjectComponentIdentifier projectComponentIdentifier = mock(ProjectComponentIdentifier.class);
+		ComponentArtifactIdentifier projectArtifactId = mock(ComponentArtifactIdentifier.class);
+		given(projectArtifactId.getComponentIdentifier()).willReturn(projectComponentIdentifier);
+		ResolvedArtifact projectArtifact = mockArtifact(fileName, group, module, version);
+		given(projectArtifact.getId()).willReturn(projectArtifactId);
+		return projectArtifact;
+	}
+
+	private ResolvedArtifact mockArtifact(String fileName, String group, String module, String version) {
+		ModuleVersionIdentifier moduleVersionIdentifier = mock(ModuleVersionIdentifier.class);
+		given(moduleVersionIdentifier.getGroup()).willReturn(group);
+		given(moduleVersionIdentifier.getName()).willReturn(module);
+		given(moduleVersionIdentifier.getVersion()).willReturn(version);
+		ResolvedModuleVersion moduleVersion = mock(ResolvedModuleVersion.class);
+		given(moduleVersion.getId()).willReturn(moduleVersionIdentifier);
+		ResolvedArtifact libraryArtifact = mock(ResolvedArtifact.class);
+		File file = new File(this.temp, fileName).getAbsoluteFile();
+		given(libraryArtifact.getFile()).willReturn(file);
+		given(libraryArtifact.getModuleVersion()).willReturn(moduleVersion);
+		return libraryArtifact;
+	}
+
+	List<String> entryLines(JarFile jarFile, String entryName) throws IOException {
+		try (BufferedReader reader = new BufferedReader(
+				new InputStreamReader(jarFile.getInputStream(jarFile.getEntry(entryName))))) {
+			return reader.lines().collect(Collectors.toList());
+		}
+	}
+
+	private Set<String> getLayerNames(List<String> index) {
+		Set<String> layerNames = new LinkedHashSet<>();
+		for (String line : index) {
+			if (line.startsWith("- ")) {
+				layerNames.add(line.substring(3, line.length() - 2));
+			}
+		}
+		return layerNames;
 	}
 
 }
