+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/test/java/org/springframework/boot/gradle/tasks/bundling/BootWarIntegrationTests.java
@@ -16,10 +16,14 @@
 
 package org.springframework.boot.gradle.tasks.bundling;
 
+import java.util.Arrays;
+import java.util.Set;
+import java.util.TreeSet;
+
 import org.springframework.boot.gradle.junit.GradleCompatibility;
 
 /**
- * Integration tests for {@link BootJar}.
+ * Integration tests for {@link BootWar}.
  *
  * @author Andy Wilkinson
  */
@@ -27,7 +31,14 @@ import org.springframework.boot.gradle.junit.GradleCompatibility;
 class BootWarIntegrationTests extends AbstractBootArchiveIntegrationTests {
 
 	BootWarIntegrationTests() {
-		super("bootWar", "WEB-INF/lib/", "WEB-INF/classes/");
+		super("bootWar", "WEB-INF/lib/", "WEB-INF/classes/", "WEB-INF/");
+	}
+
+	@Override
+	String[] getExpectedApplicationLayerContents(String... additionalFiles) {
+		Set<String> contents = new TreeSet<>(Arrays.asList(additionalFiles));
+		contents.addAll(Arrays.asList("WEB-INF/layers.idx", "META-INF/"));
+		return contents.toArray(new String[0]);
 	}
 
 }
