+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/test/resources/org/springframework/boot/gradle/tasks/bundling/BootWarIntegrationTests-developmentOnlyDependenciesAreNotIncludedInTheArchiveByDefault.gradle
@@ -16,3 +16,9 @@ dependencies {
 	developmentOnly("commons-io:commons-io:2.6")
 	implementation("commons-io:commons-io:2.6")
 }
+
+bootWar {
+	layered {
+		enabled = false
+	}
+}
