+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/test/resources/org/springframework/boot/gradle/tasks/bundling/BootWarIntegrationTests-developmentOnlyDependenciesCanBeIncludedInTheArchive.gradle
@@ -19,3 +19,9 @@ dependencies {
 bootWar {
 	classpath configurations.developmentOnly
 }
+
+bootWar {
+	layered {
+		enabled = false
+	}
+}
