+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/test/resources/org/springframework/boot/gradle/tasks/bundling/BootWarIntegrationTests-jarTypeFilteringIsApplied.gradle
@@ -17,3 +17,9 @@ dependencies {
 	implementation(name: "standard")
 	implementation(name: "starter")
 }
+
+bootWar {
+	layered {
+		enabled = false
+	}
+}
