+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/test/resources/org/springframework/boot/gradle/tasks/bundling/BootWarIntegrationTests-multiModuleImplicitLayers.gradle
@@ -1,9 +1,16 @@
 plugins {
 	id 'java'
 	id 'org.springframework.boot' version '{version}'
+	id 'war'
 }
 
-bootJar {
+subprojects {
+	apply plugin: 'java'
+	group = 'org.example.projects'
+	version = '1.2.3'
+}
+
+bootWar {
 	mainClass = 'com.example.Application'
 }
 
@@ -13,20 +20,21 @@ repositories {
 }
 
 dependencies {
+	implementation(project(':alpha'))
+	implementation(project(':bravo'))
 	implementation("com.example:library:1.0-SNAPSHOT")
 	implementation("org.apache.commons:commons-lang3:3.9")
 	implementation("org.springframework:spring-core:5.2.5.RELEASE")
-	implementation("org.springframework.boot:spring-boot-starter-logging:2.2.0.RELEASE")
 }
 
 task listLayers(type: JavaExec) {
-	classpath = bootJar.outputs.files
+	classpath = bootWar.outputs.files
 	systemProperties = [ "jarmode": "layertools" ]
 	args "list"
 }
 
 task extractLayers(type: JavaExec) {
-	classpath = bootJar.outputs.files
+	classpath = bootWar.outputs.files
 	systemProperties = [ "jarmode": "layertools" ]
 	args "extract"
 }
