+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/test/resources/org/springframework/boot/gradle/tasks/bundling/BootWarIntegrationTests-notUpToDateWhenBuiltWithoutLayersAndThenWithLayers.gradle
@@ -1,0 +1,12 @@
+plugins {
+	id 'java'
+	id 'org.springframework.boot' version '{version}'
+	id 'war'
+}
+
+bootWar {
+	mainClass = 'com.example.Application'
+	layered {
+		{layerEnablement}
+	}
+}
